Add chat utility functions
